fix: uint8 to uint64

+++ contracts/JaxProtection.sol
@@ -7,7 +7,7 @@ contract JaxProtection {
 
     struct RunProtection {
         bytes32 data_hash;
-        uint8 request_timestamp;
+        uint64 request_timestamp;
         address sender;
         bool executed;
     }
@@ -20,9 +20,9 @@ contract JaxProtection {
         RunProtection storage protection = run_protection_info[msg.sig];
         bytes32 data_hash = keccak256(msg.data);
         if(data_hash != protection.data_hash || protection.sender != msg.sender) {
-            protection.sender = msg.sender;
-            protection.data_hash = keccak256(msg.data);
-            protection.request_timestamp = uint8(block.timestamp);
+        protection.sender = msg.sender;
+        protection.data_hash = keccak256(msg.data);
+            protection.request_timestamp = uint64(block.timestamp);
             protection.executed = false;
             emit Request_Update(msg.sig, msg.data);
             return;
